--- a/model/metric.go
+++ b/model/metric.go
@@ -18,6 +18,8 @@ import (
 	"regexp"
 	"sort"
 	"strings"
+
+	"github.com/mailru/easyjson/jwriter"
 )
 
 var (
@@ -85,6 +87,29 @@ func (m Metric) Fingerprint() Fingerprint {
 // algorithm, which is, however, more susceptible to hash collisions.
 func (m Metric) FastFingerprint() Fingerprint {
 	return LabelSet(m).FastFingerprint()
+}
+
+func (m Metric) MarshalEasyJSON(w *jwriter.Writer) {
+	w.RawByte('{')
+	first := true
+	for k, v := range m {
+		if !first {
+			w.RawByte(',')
+		} else {
+			first = false
+		}
+		w.RawString(`"` + string(k) + `"`)
+		w.RawByte(':')
+		w.RawString(`"` + string(v) + `"`)
+	}
+	w.RawByte('}')
+}
+
+// MarshalJSON implements the json.Marshaler interface.
+func (m Metric) MarshalJSON() ([]byte, error) {
+	w := jwriter.Writer{}
+	m.MarshalEasyJSON(&w)
+	return w.Buffer.BuildBytes(), w.Error
 }
 
 // IsValidMetricName returns true iff name matches the pattern of MetricNameRE.