--- a/model/value.go
+++ b/model/value.go
@@ -20,6 +20,8 @@ import (
 	"sort"
 	"strconv"
 	"strings"
+
+	"github.com/mailru/easyjson/jwriter"
 )
 
 var (
@@ -42,9 +44,18 @@ var (
 // time.
 type SampleValue float64
 
+func (v SampleValue) MarshalEasyJSON(w *jwriter.Writer) {
+	w.RawByte('"')
+	w.Buffer.EnsureSpace(20)
+	w.Buffer.Buf = strconv.AppendFloat(w.Buffer.Buf, float64(v), 'f', -1, 64)
+	w.RawByte('"')
+}
+
 // MarshalJSON implements json.Marshaler.
 func (v SampleValue) MarshalJSON() ([]byte, error) {
-	return json.Marshal(v.String())
+	w := jwriter.Writer{}
+	v.MarshalEasyJSON(&w)
+	return w.Buffer.BuildBytes(), w.Error
 }
 
 // UnmarshalJSON implements json.Unmarshaler.
@@ -80,17 +91,19 @@ type SamplePair struct {
 	Value     SampleValue
 }
 
+func (s SamplePair) MarshalEasyJSON(w *jwriter.Writer) {
+	w.RawByte('[')
+	s.Timestamp.MarshalEasyJSON(w)
+	w.RawByte(',')
+	s.Value.MarshalEasyJSON(w)
+	w.RawByte(']')
+}
+
 // MarshalJSON implements json.Marshaler.
 func (s SamplePair) MarshalJSON() ([]byte, error) {
-	t, err := json.Marshal(s.Timestamp)
-	if err != nil {
-		return nil, err
-	}
-	v, err := json.Marshal(s.Value)
-	if err != nil {
-		return nil, err
-	}
-	return []byte(fmt.Sprintf("[%s,%s]", t, v)), nil
+	w := jwriter.Writer{}
+	s.MarshalEasyJSON(&w)
+	return w.Buffer.BuildBytes(), w.Error
 }
 
 // UnmarshalJSON implements json.Unmarshaler.
@@ -220,6 +233,7 @@ func (s Samples) Equal(o Samples) bool {
 }
 
 // SampleStream is a stream of Values belonging to an attached COWMetric.
+//easyjson:json
 type SampleStream struct {
 	Metric Metric       `json:"metric"`
 	Values []SamplePair `json:"values"`
@@ -303,17 +317,21 @@ type Scalar struct {
 	Timestamp Time        `json:"timestamp"`
 }
 
-func (s Scalar) String() string {
-	return fmt.Sprintf("scalar: %v @[%v]", s.Value, s.Timestamp)
+func (s Scalar) MarshalEasyJSON(w *jwriter.Writer) {
+	w.RawByte('[')
+	s.Timestamp.MarshalEasyJSON(w)
+	w.RawByte(',')
+	s.Value.MarshalEasyJSON(w)
+	w.RawByte(']')
 }
 
 // MarshalJSON implements json.Marshaler.
 func (s Scalar) MarshalJSON() ([]byte, error) {
-	v := strconv.FormatFloat(float64(s.Value), 'f', -1, 64)
-	return json.Marshal([...]interface{}{s.Timestamp, string(v)})
+	w := jwriter.Writer{}
+	s.MarshalEasyJSON(&w)
+	return w.Buffer.BuildBytes(), w.Error
 }
 
-// UnmarshalJSON implements json.Unmarshaler.
 func (s *Scalar) UnmarshalJSON(b []byte) error {
 	var f string
 	v := [...]interface{}{&s.Timestamp, &f}
@@ -328,6 +346,10 @@ func (s *Scalar) UnmarshalJSON(b []byte) error {
 	}
 	s.Value = SampleValue(value)
 	return nil
+}
+
+func (s Scalar) String() string {
+	return fmt.Sprintf("scalar: %v @[%v]", s.Value, s.Timestamp)
 }
 
 // String is a string value evaluated at the set timestamp.
@@ -353,6 +375,7 @@ func (s *String) UnmarshalJSON(b []byte) error {
 
 // Vector is basically only an alias for Samples, but the
 // contract is that in a Vector, all Samples have the same timestamp.
+//easyjson:json
 type Vector []*Sample
 
 func (vec Vector) String() string {
@@ -395,6 +418,7 @@ func (vec Vector) Equal(o Vector) bool {
 }
 
 // Matrix is a list of time series.
+//easyjson:json
 type Matrix []*SampleStream
 
 func (m Matrix) Len() int           { return len(m) }