I've been looking into options for fixing #3601 (JSON performance issues). I see that there is an open PR related to this issue #3536 -- which simply switches to jsoniter.

From my testing there are effectively 2 issues.

**First issue**
There are a few places in the marshaling process that require making lots of copies of the underlying datapoints. This not only causes memory bloat, but this significantly slows down the process. Similarly the task is made unecessarily difficult by converting int64 to float64 before putting it to a string (in the case of `Time`).

**Second Issue**
The majority of the performance issues at "large" (quoted, because large is only hours of data) are due to the large memory requirements due to marshaling the entire structure into memory before returning it. [easyjson](https://github.com/mailru/easyjson) has a streaming serialization interface which solves this problem. When using the `encoding/json` library to do the marshaling I see my responses regularly increase by ~6x in memory before being written out. With this streaming approach I see no such memory increase.

Of course any performance change is only as good as the benchmarks associated with it (which are included in this PR). On my X1 carbon 5th gen, here are the results:


**Before these changes**
```
$ go test -run=x -bench=BenchmarkMarshal
goos: linux
goarch: amd64
pkg: github.com/prometheus/common/model
BenchmarkMarshal/SampleValue/encoding/json-4         	 2000000	       617 ns/op
model.SampleValue not easyjson.Marshaler
BenchmarkMarshal/Timestamp/encoding/json-4           	 2000000	       598 ns/op
model.Time not easyjson.Marshaler
BenchmarkMarshal/SamplePair/encoding/json-4          	 1000000	      2069 ns/op
model.SamplePair not easyjson.Marshaler
BenchmarkMarshal/labelset/encoding/json-4            	 2000000	       686 ns/op
model.Metric not easyjson.Marshaler
BenchmarkMarshal/SampleStream/encoding/json-4        	   10000	    198699 ns/op
*model.SampleStream not easyjson.Marshaler
BenchmarkMarshal/Matrix/encoding/json-4              	      20	  96541250 ns/op
model.Matrix not easyjson.Marshaler
PASS
ok  	github.com/prometheus/common/model	11.913s
```


**After these changes**
```
$ go test -run=x -bench=BenchmarkMarshal
goos: linux
goarch: amd64
pkg: github.com/prometheus/common/model
BenchmarkMarshal/SampleValue/encoding/json-4         	 3000000	       516 ns/op
BenchmarkMarshal/SampleValue/easyjson-4              	10000000	       209 ns/op
BenchmarkMarshal/Timestamp/encoding/json-4           	 3000000	       518 ns/op
BenchmarkMarshal/Timestamp/easyjson-4                	10000000	       211 ns/op
BenchmarkMarshal/SamplePair/encoding/json-4          	 2000000	       719 ns/op
BenchmarkMarshal/SamplePair/easyjson-4               	 5000000	       305 ns/op
BenchmarkMarshal/labelset/encoding/json-4            	 2000000	       663 ns/op
BenchmarkMarshal/labelset/easyjson-4                 	10000000	       214 ns/op
BenchmarkMarshal/SampleStream/encoding/json-4        	   50000	     30999 ns/op
BenchmarkMarshal/SampleStream/easyjson-4             	  100000	     16747 ns/op
BenchmarkMarshal/Matrix/encoding/json-4              	     100	  14360224 ns/op
BenchmarkMarshal/Matrix/easyjson-4                   	     200	   8192013 ns/op
PASS
ok  	github.com/prometheus/common/model	25.163s
```

For ease of reading, here is the same data in a table:

|              | encoding/json (before) | encoding/json (after | encoding/json (speedup) | easyjson | easyjson (speedup) |
|--------------|------------------------|----------------------|-------------------------|----------|--------------------|
| SampleValue  | 617                    | 516                  | 16%                     | 209      | 66%                |
| Timestamp    | 598                    | 518                  | 13%                     | 211      | 64%                |
| SamplePair   | 2069                   | 719                  | 65%                     | 305      | 85%                |
| labelset     | 686                    | 663                  | 3%                      | 214      | 68%                |
| SampleStream | 198699                 | 30999                | 84%                     | 16747    | 91%                |
| Matrix       | 96541250               | 14360224             | 85%                     | 8192013  | 91%                |