@knweiss 2017-12-21
@jacksontj FWIW: You can use [benchcmp](https://godoc.org/golang.org/x/tools/cmd/benchcmp) or [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) to compare and print both benchmark results side by side.