@jacksontj 2017-12-21
@knweiss thanks, here is the output from that tool:


```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkMarshal/SampleValue/encoding/json-4      617           516           -16.37%
BenchmarkMarshal/Timestamp/encoding/json-4        598           518           -13.38%
BenchmarkMarshal/SamplePair/encoding/json-4       2069          719           -65.25%
BenchmarkMarshal/labelset/encoding/json-4         686           663           -3.35%
BenchmarkMarshal/SampleStream/encoding/json-4     198699        30999         -84.40%
BenchmarkMarshal/Matrix/encoding/json-4           96541250      14360224      -85.13%
```