@jacksontj 2017-12-22
Benmark comparison run again with memory numbers included:

```
benchmark                                         old ns/op     new ns/op     delta
BenchmarkMarshal/SampleValue/encoding/json-4      766           468           -38.90%
BenchmarkMarshal/Timestamp/encoding/json-4        713           586           -17.81%
BenchmarkMarshal/SamplePair/encoding/json-4       2050          743           -63.76%
BenchmarkMarshal/labelset/encoding/json-4         683           662           -3.07%
BenchmarkMarshal/SampleStream/encoding/json-4     197673        19165         -90.30%
BenchmarkMarshal/Matrix/encoding/json-4           96928336      8494863       -91.24%

benchmark                                         old allocs     new allocs     delta
BenchmarkMarshal/SampleValue/encoding/json-4      6              3              -50.00%
BenchmarkMarshal/Timestamp/encoding/json-4        5              4              -20.00%
BenchmarkMarshal/SamplePair/encoding/json-4       20             5              -75.00%
BenchmarkMarshal/labelset/encoding/json-4         7              7              +0.00%
BenchmarkMarshal/SampleStream/encoding/json-4     2012           118            -94.14%
BenchmarkMarshal/Matrix/encoding/json-4           1003160        50079          -95.01%

benchmark                                         old bytes     new bytes     delta
BenchmarkMarshal/SampleValue/encoding/json-4      497           400           -19.52%
BenchmarkMarshal/Timestamp/encoding/json-4        336           416           +23.81%
BenchmarkMarshal/SamplePair/encoding/json-4       1264          424           -66.46%
BenchmarkMarshal/labelset/encoding/json-4         416           416           +0.00%
BenchmarkMarshal/SampleStream/encoding/json-4     116137        5035          -95.66%
BenchmarkMarshal/Matrix/encoding/json-4           58131741      2150175       -96.30%
```