@jacksontj 2017-12-27
@juliusv Updated, I squashed the commit so the merge will be nicer :)