@bboreham 2018-01-19
You might want to take a look at #114 - a combination of the two techniques might work best.

Results from the microbenchmarks in #114 on my desktop machine:

```
BenchmarkMarshalJSON-3     	20000000	       110 ns/op	      16 B/op	       1 allocs/op
BenchmarkUnmarshalJSON-3   	20000000	        85.5 ns/op	      16 B/op	       2 allocs/op
```

and same tests against this PR:

```
BenchmarkMarshalJSON-3     	10000000	       167 ns/op	     144 B/op	       2 allocs/op
BenchmarkUnmarshalJSON-3   	10000000	       151 ns/op	      48 B/op	       2 allocs/op
```