@jacksontj 2018-01-19
@bboreham thanks for the ping! I actually have a separate branch where I've added faster marshal/unmarshal for most of the structs in common. I'm running a fork of prometheus to test out the performance gains and they are significant (>50% CPU reduction, >3x latency improvement, etc.) If this PR gets some traction I'll update it with those  (if you are interested you can see my changes -- https://github.com/prometheus/common/compare/master...jacksontj:tjackson_fork)