@brian-brazil 2018-03-21
This code isn't even being used by the Prometheus web API, and we've gone with a different json encoder.